【fix】爆款统计逻辑调整

+++ src/main/java/com/haier/hailian/contract/dto/ContractSerialDto.java
@@ -2,14 +2,16 @@ package com.haier.hailian.contract.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 合约型号汇总统计
  */
 @Data
 public class ContractSerialDto {
     private String serial;
-    private Integer yearPlan;
-    private Integer monthPlan;
-    private Integer yearSales;
-    private Integer monthSales;
+    private BigDecimal yearPlan;
+    private BigDecimal monthPlan;
+    private BigDecimal yearSales;
+    private BigDecimal monthSales;
 }
